Some work to minimize work per agent, split up some big functions

+++ src/vec2f.c
@@ -8,15 +8,15 @@ void vector2f_init(struct Vector2f *vec, float x, float y) {
 }
 
 float vector2f_length(struct Vector2f *vec) {
-  return sqrt(pow(vec->x, 2) + pow(vec->y, 2));
+  return sqrt(powf(vec->x, 2) + powf(vec->y, 2));
 }
 
 float vector2f_dist(struct Vector2f *vec1, struct Vector2f *vec2) {
-  return sqrt(pow(vec2->x - vec1->x, 2) + pow(vec2->y - vec1->y, 2));
+  return sqrt(powf(vec2->x - vec1->x, 2) + powf(vec2->y - vec1->y, 2));
 }
 
 float vector2f_dist2(struct Vector2f *vec1, struct Vector2f *vec2) {
-  return pow(vec2->x - vec1->x, 2) + pow(vec2->y - vec1->y, 2);
+  return powf(vec2->x - vec1->x, 2) + powf(vec2->y - vec1->y, 2);
 }
 
 float vector2f_angle(struct Vector2f *vec) {
